requirements.txt
ultralytics
streamlit
opencv-python-headless
numpy
scipy
lapx
pillow